should work

+++ auxilary_scripts/docker_build.sh
@@ -4,6 +4,26 @@ set -e
 
 . .env
 
+# Test if variables are set
+variable_list=("PORT" "CLUSTER_TOKEN" "EMAIL_PASSWORD" "DB_PW" "FLASK_DEBUG" "GIT_PAT_SM" "GUNICORN")
+for var_name in "${variable_list[@]}"; do
+    if [ -z "${!var_name}" ]; then
+        echo "Error: $var_name is not set in the environment."
+        exit 1
+    fi
+done
+
+if [ ! -d ".git" ]; then
+    echo "Error: This script must be run at the base of the git repository."
+    exit 1
+fi
+
+repo_name=$(git rev-parse --show-toplevel | xargs basename)
+if [ "$repo_name" != "SOM-Web" ]; then
+    echo "Error: This script must be run at the base of the git repository."
+    exit 1
+fi
+
 container_name="$(docker ps -all | grep cosmopolitan-test | awk '{print $1}')"
 if [ -n "$container_name" ]; then
     docker rm "$container_name"
@@ -25,3 +45,4 @@ docker run --name cosmopolitan-test \
     -e FLASK_DEBUG="$FLASK_DEBUG" \
     -p $PORT:$PORT \
     cosmopolitan-test
+
